results/rebuttal_new_baselines/run-20231112-040536_CodeT-SelfRefine-Reflexion-ZeroShot_donnemartin-system-design-oop-dropbox-donnemartin-system-design-oop-twitter_0_20-runs_log/dropbox/12/ZeroShot/test_file_management.py
import pytest
import file_management

def test_upload_file():
	data = {'name': 'Test', 'type': 'txt', 'size': 1, 'versions': []}
	response = file_management.upload_file(data)
	assert response['status'] == 'success'

def test_download_file():
	data = {'name': 'Test'}
	response = file_management.download_file(data)
	assert response['status'] == 'success'

def test_organize_file():
	data = {'name': 'Test', 'new_name': 'NewTest'}
	response = file_management.organize_file(data)
	assert response['status'] == 'success'

def test_version_file():
	data = {'name': 'Test'}
	response = file_management.version_file(data)
	assert response['status'] == 'success'
